Inject the triggering-signal to the command if so requested.
On cases when the instance of the triggering signal is necessary inside the Command it can now be requested it's injection by  using the [Inject(name="trigger")] tag in the Command as so:

    [Inject(name="trigger")]
    public var triggerSignal:ISignal;

This was mainly created to obtain the actual QualifiedClassName of the trigger (ISignal). But many other uses can be also applied.